api/requirements.txt
fastapi>=0.115.0
email-validator>=2.0
PySide6>=6.6
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
playwright>=1.48
APScheduler>=3.10.0
aiosqlite>=0.20.0
sqlalchemy>=2.0.0
